requirements.txt
selenium>=4.15.0
requests>=2.31.0
webdriver-manager>=4.0.0
brotli>=1.1.0
mysql-connector-python>=8.0.0
